[quantization] Factored out the default mappings
